requirements.txt
streamlit>=1.25.0
transformers>=4.31.0,<5.0.0          # For BLIP and text QA models
# PyTorch - Install from official PyTorch website
# Visit: https://pytorch.org/get-started/locally/
# For CPU: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
Pillow>=9.0.0                       # For image handling
accelerate>=0.20.0                   # For model optimization
protobuf>=3.20.0,<4.0.0              # Required by some transformers models
